requirements.txt
# Please manually install torch and torchvision
matplotlib>=3.3
numpy>=1.19
torchsummary>=1.5
tqdm>=4.45.0
efficientnet_pytorch
hpbandster
ConfigSpace